Move compressed word vector into primitives
