+++ library/Rattletrap/Json.hs
@@ -9,7 +9,7 @@ import Rattletrap.AttributeValue
 import Rattletrap.Cache
 import Rattletrap.ClassMapping
 import Rattletrap.Primitive.CompressedWord
-import Rattletrap.CompressedWordVector
+import Rattletrap.Primitive.CompressedWordVector
 import Rattletrap.Content
 import Rattletrap.Dictionary
 import Rattletrap.Float32
